Move logic to run simulation `main -> run`
